@@ -24,21 +24,30 @@
   "devDependencies": {
     "@types/dotenv": "^4.0.3",
     "@types/express": "^4.16.0",
+    "@types/express-session": "^1.15.11",
     "@types/fs-extra": "^5.0.4",
     "@types/pg": "^7.4.10",
     "@types/shelljs": "^0.8.0",
+    "eslint": "^5.6.0",
+    "eslint-config-leankit": "^4.5.0",
+    "eslint-plugin-typescript": "^0.12.0",
     "jest": "^23.6.0",
     "nodemon": "^1.18.4",
     "rimraf": "^2.6.2",
     "shelljs": "^0.8.2",
-    "ts-jest": "^23.1.4",
+    "ts-jest": "^23.10.2",
     "ts-node": "^7.0.1",
-    "typescript": "^3.0.3"
+    "typescript": "^3.1.1",
+    "typescript-eslint-parser": "^19.0.1"
   },
   "dependencies": {
+    "@okta/jwt-verifier": "0.0.12",
+    "@okta/oidc-middleware": "^1.0.0",
+    "@okta/okta-sdk-nodejs": "^1.2.0",
     "body-parser": "^1.18.3",
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
+    "express-session": "^1.15.6",
     "fs-extra": "^7.0.0",
     "pg-monitor": "^1.0.0",
     "pg-promise": "^8.4.6"
