@@ -1,32 +1,32 @@
-import fs = require("fs");
-import os = require("os");
+import fs from "fs";
+import os from "os";
 import * as pgMonitor from "pg-monitor";
-import {IOptions} from "pg-promise";
+import { IOptions } from "pg-promise";
 
-pgMonitor.setTheme("matrix");
+pgMonitor.setTheme( "matrix" );
 const $DEV = process.env.NODE_ENV === "development";
 const logFile = "./db-errors.log";
 
-pgMonitor.setLog((msg, info) => {
-    if (info.event === "error") {
-        let logText = os.EOL + msg;
-        if (info.time) {
-            logText = os.EOL + logText;
-        }
-        fs.appendFileSync(logFile, logText);
-    }
+pgMonitor.setLog( ( msg, info ) => {
+	if ( info.event === "error" ) {
+		let logText = os.EOL + msg;
+		if ( info.time ) {
+			logText = os.EOL + logText;
+		}
+		fs.appendFileSync( logFile, logText );
+	}
 
-    if (!$DEV) {
-        info.display = false;
-    }
-});
+	if ( !$DEV ) {
+		info.display = false;
+	}
+} );
 
 export = {
-    init(options: IOptions<any>) {
-        if ($DEV) {
-            pgMonitor.attach(options);
-        } else {
-            pgMonitor.attach(options, ["error"]);
-        }
-    },
+	init( options: IOptions<any> ) {
+		if ( $DEV ) {
+			pgMonitor.attach( options );
+		} else {
+			pgMonitor.attach( options, [ "error" ] );
+		}
+	}
 };
