@@ -1,8 +1,23 @@
 const register = async ( { db, api }: { db: any, api: any } ) => {
-    api.get("/guitars/create", () => db.guitars.create());
-    api.get("/guitars/find/:userId/:name", (req: any) => db.guitars.find(req.params));
-    api.get("/guitars/all/:userId", (req: any) => db.guitars.all(req.params));
-    api.get("/guitars/total/:userId", (req: any) => db.guitars.total(req.params));
+	api.post( "/guitars/add", ( req: any, userId: string ) => {
+		return db.guitars.create( { userId, ...req.body } );
+	} );
+
+	api.remove( "/guitars/remove/:id", ( req: any, userId: string ) => {
+		return db.guitars.remove( { userId, ...req.params } );
+	} );
+
+	api.get( "/guitars/find/:search", ( req: any, userId: string ) => {
+		return db.guitars.find( { userId, ...req.params } );
+	} );
+
+	api.get( "/guitars/all", ( req: any, userId: string ) => {
+		return db.guitars.all( { userId } );
+	} );
+
+	api.get( "/guitars/total", ( req: any, userId: string ) => {
+		return db.guitars.total( { userId } );
+	} );
 };
 
 export { register };
