@@ -1,3 +1,3 @@
 import * as shell from "shelljs";
 
-shell.cp("-R", "src/db/sql", "dist/db/");
+shell.cp( "-R", "src/db/sql", "dist/db/" );
